Check runtime radix bit intervals before narrowing
Preserve runtime bit bounds as signed 64-bit values until the wrapper
checks their interval against the key width. Trap invalid intervals
before converting to CUB integer arguments.

+++ python/cuda_coop/tests/contracts/core/test_core_block_radix_sort.py
@@ -6,7 +6,7 @@
 import numpy as np
 import pytest
 
-from cuda.coop._core import INT32, Array, Dependency, TempStorageParameter, Value
+from cuda.coop._core import INT64, Array, Dependency, TempStorageParameter, Value
 from cuda.coop._core.block import (
     BlockRadixSortBitPolicy,
     BlockRadixSortOutput,
@@ -93,8 +93,8 @@ def test_explicit_descending_striped_pair_sort_owns_runtime_bit_abi():
                 is_inout=True,
                 is_return=False,
             ),
-            Value(INT32, name="begin_bit"),
-            Value(INT32, name="end_bit"),
+            Value(INT64, name="begin_bit"),
+            Value(INT64, name="end_bit"),
         ),
     )
